update files to conform MVC
